refactor(store): split monolithic sqlite.go into domain-specific files
- Extract user operations into user.go (UserReader + UserWriter)
- Extract client operations into client.go (ClientReader + ClientWriter)
- Extract token operations into token.go (TokenReader + TokenWriter)
- Extract device code operations into device_code.go (DeviceCodeStore)
- Extract authorization code operations into authorization_code.go
- Extract user authorization operations into user_authorization.go
- Extract OAuth connection operations into oauth_connection.go
- Extract audit log operations into audit_log.go (AuditStore)
- Extract cleanup and metrics operations into cleanup.go
- Keep Store struct, New(), seedData(), and infrastructure methods in sqlite.go
