+++ internal/store/sqlite.go
@@ -8,7 +8,6 @@ import (
 	"log"
 	"os"
 	"strings"
-	"time"
 
 	"github.com/go-authgate/authgate/internal/config"
 	"github.com/go-authgate/authgate/internal/core"
@@ -18,7 +17,6 @@ import (
 	"github.com/google/uuid"
 	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
 	"gorm.io/gorm/logger"
 )
 
@@ -267,389 +265,6 @@ func (s *Store) seedData(ctx context.Context, cfg *config.Config) error {
 	return nil
 }
 
-// User operations
-func (s *Store) GetUserByUsername(username string) (*models.User, error) {
-	var user models.User
-	if err := s.db.Where("username = ?", username).First(&user).Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-func (s *Store) GetUserByID(id string) (*models.User, error) {
-	var user models.User
-	if err := s.db.Where("id = ?", id).First(&user).Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-// GetUserByExternalID finds a user by their external ID and auth source
-func (s *Store) GetUserByExternalID(externalID, authSource string) (*models.User, error) {
-	var user models.User
-	if err := s.db.Where("external_id = ? AND auth_source = ?", externalID, authSource).
-		First(&user).
-		Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-// UpsertExternalUser creates or updates a user from external authentication
-func (s *Store) UpsertExternalUser(
-	username, externalID, authSource, email, fullName string,
-) (*models.User, error) {
-	var user models.User
-
-	// Try to find existing user by external ID
-	err := s.db.Where("external_id = ? AND auth_source = ?", externalID, authSource).
-		First(&user).
-		Error
-
-	if err == nil {
-		// User exists - check if username changed
-		if user.Username != username {
-			// Username changed, verify new username is available
-			var conflictingUser models.User
-			conflictErr := s.db.Where("username = ? AND id != ?", username, user.ID).
-				First(&conflictingUser).
-				Error
-
-			if conflictErr == nil {
-				// Username taken by another user
-				return nil, ErrUsernameConflict
-			}
-			if !errors.Is(conflictErr, gorm.ErrRecordNotFound) {
-				// Unexpected database error
-				return nil, fmt.Errorf("failed to check username: %w", conflictErr)
-			}
-			// Username available, continue with update
-		}
-
-		// Update user fields
-		user.Username = username
-		user.Email = email
-		user.FullName = fullName
-		if err := s.db.Save(&user).Error; err != nil {
-			return nil, fmt.Errorf("failed to update external user: %w", err)
-		}
-		return &user, nil
-	}
-
-	// Handle query error
-	if !errors.Is(err, gorm.ErrRecordNotFound) {
-		return nil, fmt.Errorf("failed to query external user: %w", err)
-	}
-
-	// User doesn't exist - check if username is available
-	var existingUser models.User
-	err = s.db.Where("username = ?", username).First(&existingUser).Error
-
-	if err == nil {
-		// Username already taken
-		return nil, ErrUsernameConflict
-	}
-	if !errors.Is(err, gorm.ErrRecordNotFound) {
-		// Unexpected database error
-		return nil, fmt.Errorf("failed to check username: %w", err)
-	}
-
-	// Create new user
-	user = models.User{
-		ID:           uuid.New().String(),
-		Username:     username,
-		PasswordHash: "", // No local password for external users
-		Role:         models.UserRoleUser,
-		ExternalID:   externalID,
-		AuthSource:   authSource,
-		Email:        email,
-		FullName:     fullName,
-	}
-
-	if err := s.db.Create(&user).Error; err != nil {
-		return nil, fmt.Errorf("failed to create external user: %w", err)
-	}
-
-	return &user, nil
-}
-
-// OAuth Client operations
-func (s *Store) GetClient(clientID string) (*models.OAuthApplication, error) {
-	var client models.OAuthApplication
-	if err := s.db.Where("client_id = ?", clientID).First(&client).Error; err != nil {
-		return nil, err
-	}
-	return &client, nil
-}
-
-// ListClientsPaginated returns paginated OAuth clients with search and optional status filter support
-func (s *Store) ListClientsPaginated(
-	params PaginationParams,
-) ([]models.OAuthApplication, PaginationResult, error) {
-	var clients []models.OAuthApplication
-	var total int64
-
-	// Build base query
-	query := s.db.Model(&models.OAuthApplication{})
-
-	// Apply search filter if provided
-	if params.Search != "" {
-		searchPattern := "%" + params.Search + "%"
-		query = query.Where(
-			"client_name LIKE ? OR client_id LIKE ? OR description LIKE ?",
-			searchPattern, searchPattern, searchPattern,
-		)
-	}
-
-	// Apply status filter if provided
-	if params.StatusFilter != "" {
-		query = query.Where("status = ?", params.StatusFilter)
-	}
-
-	// Count total records
-	if err := query.Count(&total).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	// Calculate pagination
-	pagination := CalculatePagination(total, params.Page, params.PageSize)
-
-	// Apply pagination and fetch results
-	if err := query.Order("created_at DESC").
-		Limit(params.PageSize).
-		Offset(pagination.Offset()).
-		Find(&clients).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	return clients, pagination, nil
-}
-
-// ListClientsByUserID returns paginated OAuth clients owned by the given user
-func (s *Store) ListClientsByUserID(
-	userID string,
-	params PaginationParams,
-) ([]models.OAuthApplication, PaginationResult, error) {
-	var clients []models.OAuthApplication
-	var total int64
-
-	query := s.db.Model(&models.OAuthApplication{}).Where("user_id = ?", userID)
-
-	if params.Search != "" {
-		searchPattern := "%" + params.Search + "%"
-		query = query.Where(
-			"client_name LIKE ? OR client_id LIKE ? OR description LIKE ?",
-			searchPattern, searchPattern, searchPattern,
-		)
-	}
-
-	if err := query.Count(&total).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	pagination := CalculatePagination(total, params.Page, params.PageSize)
-
-	if err := query.Order("created_at DESC").
-		Limit(params.PageSize).
-		Offset(pagination.Offset()).
-		Find(&clients).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	return clients, pagination, nil
-}
-
-// CountClientsByStatus returns the number of clients with the given status
-func (s *Store) CountClientsByStatus(status string) (int64, error) {
-	var count int64
-	err := s.db.Model(&models.OAuthApplication{}).
-		Where("status = ?", status).
-		Count(&count).Error
-	return count, err
-}
-
-func (s *Store) GetClientsByIDs(clientIDs []string) (map[string]*models.OAuthApplication, error) {
-	if len(clientIDs) == 0 {
-		return make(map[string]*models.OAuthApplication), nil
-	}
-
-	var clients []models.OAuthApplication
-	if err := s.db.Where("client_id IN ?", clientIDs).Find(&clients).Error; err != nil {
-		return nil, err
-	}
-
-	// Convert to map for easy lookup
-	clientMap := make(map[string]*models.OAuthApplication, len(clients))
-	for i := range clients {
-		clientMap[clients[i].ClientID] = &clients[i]
-	}
-
-	return clientMap, nil
-}
-
-// GetUsersByIDs batch loads users by IDs using WHERE IN to prevent N+1 queries
-func (s *Store) GetUsersByIDs(userIDs []string) (map[string]*models.User, error) {
-	if len(userIDs) == 0 {
-		return make(map[string]*models.User), nil
-	}
-
-	var users []models.User
-	if err := s.db.Where("id IN ?", userIDs).Find(&users).Error; err != nil {
-		return nil, err
-	}
-
-	// Convert to map for O(1) lookup
-	userMap := make(map[string]*models.User, len(users))
-	for i := range users {
-		userMap[users[i].ID] = &users[i]
-	}
-
-	return userMap, nil
-}
-
-func (s *Store) CreateClient(client *models.OAuthApplication) error {
-	return s.db.Create(client).Error
-}
-
-func (s *Store) UpdateClient(client *models.OAuthApplication) error {
-	return s.db.Save(client).Error
-}
-
-func (s *Store) DeleteClient(clientID string) error {
-	return s.db.Where("client_id = ?", clientID).Delete(&models.OAuthApplication{}).Error
-}
-
-// Device Code operations
-
-// CreateDeviceCode creates a new device code
-func (s *Store) CreateDeviceCode(dc *models.DeviceCode) error {
-	return s.db.Create(dc).Error
-}
-
-// GetDeviceCodesByID retrieves all device codes with matching ID suffix
-// Used for hash verification during token exchange
-func (s *Store) GetDeviceCodesByID(deviceCodeID string) ([]*models.DeviceCode, error) {
-	var dcs []*models.DeviceCode
-	if err := s.db.Where("device_code_id = ?", deviceCodeID).Find(&dcs).Error; err != nil {
-		return nil, err
-	}
-	return dcs, nil
-}
-
-// GetDeviceCodeByUserCode retrieves a device code by user code
-func (s *Store) GetDeviceCodeByUserCode(userCode string) (*models.DeviceCode, error) {
-	var dc models.DeviceCode
-	if err := s.db.Where("user_code = ?", userCode).First(&dc).Error; err != nil {
-		return nil, err
-	}
-	return &dc, nil
-}
-
-// UpdateDeviceCode updates a device code
-func (s *Store) UpdateDeviceCode(dc *models.DeviceCode) error {
-	return s.db.Save(dc).Error
-}
-
-// DeleteDeviceCodeByID deletes device code by ID (primary key)
-func (s *Store) DeleteDeviceCodeByID(id int64) error {
-	return s.db.Delete(&models.DeviceCode{}, id).Error
-}
-
-// Access Token operations
-func (s *Store) CreateAccessToken(token *models.AccessToken) error {
-	return s.db.Create(token).Error
-}
-
-func (s *Store) GetAccessTokenByHash(hash string) (*models.AccessToken, error) {
-	var t models.AccessToken
-	if err := s.db.Where("token_hash = ?", hash).First(&t).Error; err != nil {
-		return nil, err
-	}
-	return &t, nil
-}
-
-func (s *Store) GetAccessTokenByID(tokenID string) (*models.AccessToken, error) {
-	var t models.AccessToken
-	if err := s.db.Where("id = ?", tokenID).First(&t).Error; err != nil {
-		return nil, err
-	}
-	return &t, nil
-}
-
-func (s *Store) GetTokensByUserID(userID string) ([]models.AccessToken, error) {
-	var tokens []models.AccessToken
-	if err := s.db.Where("user_id = ?", userID).
-		Order("created_at DESC").
-		Find(&tokens).Error; err != nil {
-		return nil, err
-	}
-	return tokens, nil
-}
-
-// GetTokensByUserIDPaginated returns paginated tokens for a user with search support
-func (s *Store) GetTokensByUserIDPaginated(
-	userID string,
-	params PaginationParams,
-) ([]models.AccessToken, PaginationResult, error) {
-	var tokens []models.AccessToken
-	var total int64
-
-	// Build base query
-	query := s.db.Model(&models.AccessToken{}).Where("user_id = ?", userID)
-
-	// Apply search filter if provided (search in scopes or join with clients for client_name)
-	if params.Search != "" {
-		searchPattern := "%" + params.Search + "%"
-		// Search in scopes or client_id
-		// For client_name search, we'll need to join with oauth_applications
-		query = query.Where(
-			"scopes LIKE ? OR client_id IN (?)",
-			searchPattern,
-			s.db.Model(&models.OAuthApplication{}).
-				Select("client_id").
-				Where("client_name LIKE ?", searchPattern),
-		)
-	}
-
-	// Count total records
-	if err := query.Count(&total).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	// Calculate pagination
-	pagination := CalculatePagination(total, params.Page, params.PageSize)
-
-	// Apply pagination and fetch results
-	if err := query.Order("created_at DESC").
-		Limit(params.PageSize).
-		Offset(pagination.Offset()).
-		Find(&tokens).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	return tokens, pagination, nil
-}
-
-func (s *Store) RevokeToken(tokenID string) error {
-	return s.db.Where("id = ?", tokenID).Delete(&models.AccessToken{}).Error
-}
-
-func (s *Store) RevokeTokensByUserID(userID string) error {
-	return s.db.Where("user_id = ?", userID).Delete(&models.AccessToken{}).Error
-}
-
-func (s *Store) RevokeTokensByClientID(clientID string) error {
-	return s.db.Where("client_id = ?", clientID).Delete(&models.AccessToken{}).Error
-}
-
-func (s *Store) DeleteExpiredTokens() error {
-	return s.db.Where("expires_at < ?", time.Now()).Delete(&models.AccessToken{}).Error
-}
-
-func (s *Store) DeleteExpiredDeviceCodes() error {
-	return s.db.Where("expires_at < ?", time.Now()).Delete(&models.DeviceCode{}).Error
-}
-
 // Health checks the database connection
 func (s *Store) Health() error {
 	sqlDB, err := s.db.DB()
@@ -672,467 +287,4 @@ func (s *Store) RunInTransaction(fn func(tx core.Store) error) error {
 	return s.db.Transaction(func(tx *gorm.DB) error {
 		return fn(&Store{db: tx})
 	})
-}
-
-// UpdateTokenStatus updates the status of a token
-func (s *Store) UpdateTokenStatus(tokenID, status string) error {
-	return s.db.Model(&models.AccessToken{}).
-		Where("id = ?", tokenID).
-		Update("status", status).Error
-}
-
-// UpdateTokenLastUsedAt updates the last_used_at timestamp of a token
-func (s *Store) UpdateTokenLastUsedAt(tokenID string, t time.Time) error {
-	return s.db.Model(&models.AccessToken{}).
-		Where("id = ?", tokenID).
-		Update("last_used_at", &t).Error
-}
-
-// RevokeTokenFamily revokes all active tokens that share the same TokenFamilyID.
-// This is used for refresh token rotation replay detection: when a revoked refresh token
-// is reused, all tokens in the family must be invalidated to prevent stolen token abuse.
-func (s *Store) RevokeTokenFamily(familyID string) (int64, error) {
-	result := s.db.Model(&models.AccessToken{}).
-		Where("token_family_id = ? AND status = ?", familyID, models.TokenStatusActive).
-		Update("status", models.TokenStatusRevoked)
-	return result.RowsAffected, result.Error
-}
-
-// GetTokensByCategoryAndStatus returns tokens filtered by category and status
-func (s *Store) GetTokensByCategoryAndStatus(
-	userID, category, status string,
-) ([]models.AccessToken, error) {
-	var tokens []models.AccessToken
-	err := s.db.Where("user_id = ? AND token_category = ? AND status = ?", userID, category, status).
-		Order("created_at DESC").
-		Find(&tokens).
-		Error
-	return tokens, err
-}
-
-// OAuth Connection operations
-
-// GetUserByEmail finds a user by email address
-func (s *Store) GetUserByEmail(email string) (*models.User, error) {
-	var user models.User
-	if err := s.db.Where("email = ?", email).First(&user).Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-// CreateUser creates a new user
-func (s *Store) CreateUser(user *models.User) error {
-	return s.db.Create(user).Error
-}
-
-// UpdateUser updates an existing user
-func (s *Store) UpdateUser(user *models.User) error {
-	return s.db.Save(user).Error
-}
-
-// DeleteUser deletes a user by ID
-func (s *Store) DeleteUser(id string) error {
-	return s.db.Delete(&models.User{}, "id = ?", id).Error
-}
-
-// CreateOAuthConnection creates a new OAuth connection
-func (s *Store) CreateOAuthConnection(conn *models.OAuthConnection) error {
-	return s.db.Create(conn).Error
-}
-
-// GetOAuthConnection finds an OAuth connection by provider and provider user ID
-func (s *Store) GetOAuthConnection(
-	provider, providerUserID string,
-) (*models.OAuthConnection, error) {
-	var conn models.OAuthConnection
-	err := s.db.Where("provider = ? AND provider_user_id = ?", provider, providerUserID).
-		First(&conn).Error
-	if err != nil {
-		return nil, err
-	}
-	return &conn, nil
-}
-
-// GetOAuthConnectionByUserAndProvider finds an OAuth connection by user ID and provider
-func (s *Store) GetOAuthConnectionByUserAndProvider(
-	userID, provider string,
-) (*models.OAuthConnection, error) {
-	var conn models.OAuthConnection
-	err := s.db.Where("user_id = ? AND provider = ?", userID, provider).
-		First(&conn).Error
-	if err != nil {
-		return nil, err
-	}
-	return &conn, nil
-}
-
-// GetOAuthConnectionsByUserID returns all OAuth connections for a user
-func (s *Store) GetOAuthConnectionsByUserID(userID string) ([]models.OAuthConnection, error) {
-	var conns []models.OAuthConnection
-	err := s.db.Where("user_id = ?", userID).
-		Order("created_at DESC").
-		Find(&conns).Error
-	return conns, err
-}
-
-// UpdateOAuthConnection updates an existing OAuth connection
-func (s *Store) UpdateOAuthConnection(conn *models.OAuthConnection) error {
-	return s.db.Save(conn).Error
-}
-
-// DeleteOAuthConnection deletes an OAuth connection by ID
-func (s *Store) DeleteOAuthConnection(id string) error {
-	return s.db.Delete(&models.OAuthConnection{}, "id = ?", id).Error
-}
-
-// Audit Log operations
-
-// CreateAuditLog creates a single audit log entry
-func (s *Store) CreateAuditLog(log *models.AuditLog) error {
-	return s.db.Create(log).Error
-}
-
-// CreateAuditLogBatch creates multiple audit log entries in a single transaction
-func (s *Store) CreateAuditLogBatch(logs []*models.AuditLog) error {
-	if len(logs) == 0 {
-		return nil
-	}
-	return s.db.CreateInBatches(logs, 100).Error
-}
-
-// GetAuditLogsPaginated retrieves audit logs with pagination and filtering
-func (s *Store) GetAuditLogsPaginated(
-	params PaginationParams,
-	filters AuditLogFilters,
-) ([]models.AuditLog, PaginationResult, error) {
-	var logs []models.AuditLog
-	var total int64
-
-	// Build base query
-	query := s.db.Model(&models.AuditLog{})
-
-	// Apply filters
-	if filters.EventType != "" {
-		query = query.Where("event_type = ?", filters.EventType)
-	}
-	if filters.ActorUserID != "" {
-		query = query.Where("actor_user_id = ?", filters.ActorUserID)
-	}
-	if filters.ResourceType != "" {
-		query = query.Where("resource_type = ?", filters.ResourceType)
-	}
-	if filters.ResourceID != "" {
-		query = query.Where("resource_id = ?", filters.ResourceID)
-	}
-	if filters.Severity != "" {
-		query = query.Where("severity = ?", filters.Severity)
-	}
-	if filters.Success != nil {
-		query = query.Where("success = ?", *filters.Success)
-	}
-	if !filters.StartTime.IsZero() {
-		query = query.Where("event_time >= ?", filters.StartTime)
-	}
-	if !filters.EndTime.IsZero() {
-		query = query.Where("event_time <= ?", filters.EndTime)
-	}
-	if filters.ActorIP != "" {
-		query = query.Where("actor_ip = ?", filters.ActorIP)
-	}
-	if filters.Search != "" {
-		searchPattern := "%" + filters.Search + "%"
-		query = query.Where(
-			"action LIKE ? OR resource_name LIKE ? OR actor_username LIKE ?",
-			searchPattern, searchPattern, searchPattern,
-		)
-	}
-
-	// Count total records
-	if err := query.Count(&total).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	// Calculate pagination
-	pagination := CalculatePagination(total, params.Page, params.PageSize)
-
-	// Apply pagination and fetch results
-	if err := query.Order("event_time DESC").
-		Limit(params.PageSize).
-		Offset(pagination.Offset()).
-		Find(&logs).Error; err != nil {
-		return nil, PaginationResult{}, err
-	}
-
-	return logs, pagination, nil
-}
-
-// DeleteOldAuditLogs deletes audit logs older than the specified time
-func (s *Store) DeleteOldAuditLogs(olderThan time.Time) (int64, error) {
-	result := s.db.Where("created_at < ?", olderThan).Delete(&models.AuditLog{})
-	return result.RowsAffected, result.Error
-}
-
-// GetAuditLogStats returns statistics about audit logs in a given time range
-func (s *Store) GetAuditLogStats(startTime, endTime time.Time) (AuditLogStats, error) {
-	stats := AuditLogStats{
-		EventsByType:     make(map[models.EventType]int64),
-		EventsBySeverity: make(map[models.EventSeverity]int64),
-	}
-
-	// Build base query as a function to avoid GORM query state mutation
-	baseQuery := func() *gorm.DB {
-		q := s.db.Model(&models.AuditLog{})
-		if !startTime.IsZero() {
-			q = q.Where("event_time >= ?", startTime)
-		}
-		if !endTime.IsZero() {
-			q = q.Where("event_time <= ?", endTime)
-		}
-		return q
-	}
-
-	// Total events
-	if err := baseQuery().Count(&stats.TotalEvents).Error; err != nil {
-		return stats, err
-	}
-
-	// Success/Failure counts
-	if err := baseQuery().Where("success = ?", true).Count(&stats.SuccessCount).Error; err != nil {
-		return stats, err
-	}
-	stats.FailureCount = stats.TotalEvents - stats.SuccessCount
-
-	// Events by type
-	var typeResults []struct {
-		EventType models.EventType
-		Count     int64
-	}
-	if err := baseQuery().Select("event_type, COUNT(*) as count").
-		Group("event_type").
-		Find(&typeResults).Error; err != nil {
-		return stats, err
-	}
-	for _, r := range typeResults {
-		stats.EventsByType[r.EventType] = r.Count
-	}
-
-	// Events by severity
-	var severityResults []struct {
-		Severity models.EventSeverity
-		Count    int64
-	}
-	if err := baseQuery().Select("severity, COUNT(*) as count").
-		Group("severity").
-		Find(&severityResults).Error; err != nil {
-		return stats, err
-	}
-	for _, r := range severityResults {
-		stats.EventsBySeverity[r.Severity] = r.Count
-	}
-
-	return stats, nil
-}
-
-// CountActiveTokensByCategory counts active, non-expired tokens by category
-func (s *Store) CountActiveTokensByCategory(category string) (int64, error) {
-	var count int64
-	err := s.db.Model(&models.AccessToken{}).
-		Where("token_category = ? AND status = ? AND expires_at > ?",
-			category, models.TokenStatusActive, time.Now()).
-		Count(&count).
-		Error
-	return count, err
-}
-
-// CountTotalDeviceCodes counts all non-expired device codes
-func (s *Store) CountTotalDeviceCodes() (int64, error) {
-	var count int64
-	err := s.db.Model(&models.DeviceCode{}).
-		Where("expires_at > ?", time.Now()).
-		Count(&count).
-		Error
-	return count, err
-}
-
-// CountPendingDeviceCodes counts pending (not yet authorized) device codes
-func (s *Store) CountPendingDeviceCodes() (int64, error) {
-	var count int64
-	err := s.db.Model(&models.DeviceCode{}).
-		Where("expires_at > ? AND authorized = ?", time.Now(), false).
-		Count(&count).
-		Error
-	return count, err
-}
-
-// ============================================================
-// Authorization Code operations
-// ============================================================
-
-// GetClientByIntID retrieves an OAuth application by its integer primary key
-func (s *Store) GetClientByIntID(id int64) (*models.OAuthApplication, error) {
-	var client models.OAuthApplication
-	if err := s.db.Where("id = ?", id).First(&client).Error; err != nil {
-		return nil, err
-	}
-	return &client, nil
-}
-
-// CreateAuthorizationCode persists a new authorization code
-func (s *Store) CreateAuthorizationCode(code *models.AuthorizationCode) error {
-	return s.db.Create(code).Error
-}
-
-// GetAuthorizationCodeByHash retrieves an authorization code by its SHA-256 hash
-func (s *Store) GetAuthorizationCodeByHash(hash string) (*models.AuthorizationCode, error) {
-	var code models.AuthorizationCode
-	if err := s.db.Where("code_hash = ?", hash).First(&code).Error; err != nil {
-		return nil, err
-	}
-	return &code, nil
-}
-
-// MarkAuthorizationCodeUsed atomically sets UsedAt only when the code has not
-// yet been consumed.  The WHERE clause includes "used_at IS NULL" so that a
-// concurrent request that races past the application-level IsUsed() check will
-// update 0 rows and receive ErrAuthCodeAlreadyUsed, preventing double issuance.
-func (s *Store) MarkAuthorizationCodeUsed(id uint) error {
-	now := time.Now()
-	result := s.db.Model(&models.AuthorizationCode{}).
-		Where("id = ? AND used_at IS NULL", id).
-		Update("used_at", &now)
-	if result.Error != nil {
-		return result.Error
-	}
-	if result.RowsAffected == 0 {
-		return ErrAuthCodeAlreadyUsed
-	}
-	return nil
-}
-
-// ============================================================
-// UserAuthorization operations
-// ============================================================
-
-// GetUserAuthorization retrieves the active consent record for a (user, application) pair
-func (s *Store) GetUserAuthorization(
-	userID string,
-	applicationID int64,
-) (*models.UserAuthorization, error) {
-	var auth models.UserAuthorization
-	if err := s.db.Where("user_id = ? AND application_id = ? AND is_active = ?", userID, applicationID, true).
-		First(&auth).
-		Error; err != nil {
-		return nil, err
-	}
-	return &auth, nil
-}
-
-// GetUserAuthorizationByUUID retrieves an authorization by its public UUID, scoped to the owner
-func (s *Store) GetUserAuthorizationByUUID(
-	authUUID, userID string,
-) (*models.UserAuthorization, error) {
-	var auth models.UserAuthorization
-	if err := s.db.Where("uuid = ? AND user_id = ?", authUUID, userID).
-		First(&auth).
-		Error; err != nil {
-		return nil, err
-	}
-	return &auth, nil
-}
-
-// UpsertUserAuthorization creates a new consent record or re-activates and updates an existing one.
-// Uses a single atomic INSERT ... ON CONFLICT DO UPDATE to avoid the race condition that arises
-// from a non-atomic SELECT-then-INSERT/UPDATE pattern.
-func (s *Store) UpsertUserAuthorization(auth *models.UserAuthorization) error {
-	now := time.Now()
-	auth.GrantedAt = now
-	auth.IsActive = true
-	auth.RevokedAt = nil
-	return s.db.Clauses(clause.OnConflict{
-		Columns: []clause.Column{
-			{Name: "user_id"},
-			{Name: "application_id"},
-		},
-		DoUpdates: clause.AssignmentColumns([]string{
-			"uuid", "client_id", "scopes", "granted_at", "revoked_at", "is_active", "updated_at",
-		}),
-	}).Create(auth).Error
-}
-
-// RevokeUserAuthorization marks an authorization as revoked and returns the record
-func (s *Store) RevokeUserAuthorization(
-	authUUID, userID string,
-) (*models.UserAuthorization, error) {
-	var auth models.UserAuthorization
-	if err := s.db.Where("uuid = ? AND user_id = ?", authUUID, userID).
-		First(&auth).
-		Error; err != nil {
-		return nil, err
-	}
-	now := time.Now()
-	if err := s.db.Model(&auth).Updates(map[string]any{
-		"is_active":  false,
-		"revoked_at": &now,
-	}).Error; err != nil {
-		return nil, err
-	}
-	return &auth, nil
-}
-
-// ListUserAuthorizations returns all active authorizations for a user, newest first
-func (s *Store) ListUserAuthorizations(userID string) ([]models.UserAuthorization, error) {
-	var auths []models.UserAuthorization
-	if err := s.db.Where("user_id = ? AND is_active = ?", userID, true).
-		Order("granted_at DESC").
-		Find(&auths).Error; err != nil {
-		return nil, err
-	}
-	return auths, nil
-}
-
-// RevokeTokensByAuthorizationID revokes all active tokens linked to a specific UserAuthorization
-func (s *Store) RevokeTokensByAuthorizationID(authorizationID uint) error {
-	return s.db.Model(&models.AccessToken{}).
-		Where("authorization_id = ? AND status = ?", authorizationID, models.TokenStatusActive).
-		Update("status", models.TokenStatusRevoked).Error
-}
-
-// RevokeAllActiveTokensByClientID revokes every active token for a client and returns the count
-func (s *Store) RevokeAllActiveTokensByClientID(clientID string) (int64, error) {
-	result := s.db.Model(&models.AccessToken{}).
-		Where("client_id = ? AND status = ?", clientID, models.TokenStatusActive).
-		Update("status", models.TokenStatusRevoked)
-	return result.RowsAffected, result.Error
-}
-
-// RevokeAllUserAuthorizationsByClientID invalidates all active consent records for a client
-func (s *Store) RevokeAllUserAuthorizationsByClientID(clientID string) error {
-	now := time.Now()
-	return s.db.Model(&models.UserAuthorization{}).
-		Where("client_id = ? AND is_active = ?", clientID, true).
-		Updates(map[string]any{
-			"is_active":  false,
-			"revoked_at": &now,
-		}).Error
-}
-
-// CountActiveTokensByClientID counts active tokens for a specific client
-func (s *Store) CountActiveTokensByClientID(clientID string) (int64, error) {
-	var count int64
-	err := s.db.Model(&models.AccessToken{}).
-		Where("client_id = ? AND status = ?", clientID, models.TokenStatusActive).
-		Count(&count).Error
-	return count, err
-}
-
-// GetClientAuthorizations returns all active consent records for a client, ordered by grant date
-func (s *Store) GetClientAuthorizations(clientID string) ([]models.UserAuthorization, error) {
-	var auths []models.UserAuthorization
-	if err := s.db.Where("client_id = ? AND is_active = ?", clientID, true).
-		Order("granted_at DESC").
-		Find(&auths).Error; err != nil {
-		return nil, err
-	}
-	return auths, nil
 }
